+++ src/systems/si/units/mod.rs
@@ -7,11 +7,12 @@ mod scaled;
 pub use base::{Ampere, Candela, Kelvin, Kilogram, Meter, Mole, Second};
 pub use derived::{
     CubicMeter, Gray, Hertz, Joule, JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole,
-    JoulePerMoleKelvin, KilogramPerCubicMeter, MeterPerSecond, Pascal, PerKelvin, PerSecond,
-    PerSquareKelvin, SquareMeter, SquareMeterPerSecond, Watt, WattPerMeterKelvin,
+    JoulePerMoleKelvin, JoulePerSquareMeter, KilogramPerCubicMeter, MeterPerSecond, Pascal,
+    PerKelvin, PerMeter, PerSecond, PerSquareKelvin, SquareMeter, SquareMeterPerKilogram,
+    SquareMeterPerSecond, Watt, WattPerMeterKelvin,
 };
 pub use scaled::{
     Centimeter, CubicMillimeter, ElectronVolt, Gram, GramPerCubicCentimeter, Kilohertz, Kilometer,
     Kilopascal, MegaElectronVolt, Megahertz, Megapascal, Microsecond, Millimeter, Millisecond,
-    SquareCentimeter,
+    PerCentimeter, SquareCentimeter, SquareCentimeterPerGram,
 };